+++ packages/adminui-vue/components/navbarSlotDefault/UNavbarSearchButton.vue
@@ -175,9 +175,12 @@ export default {
         }
       }]
 
+      let qText = this.query
+      if (qText[0] === '№') qText = '"' + qText + '"' // do not split № on lexemas - search as is
+
       let repo = this.$UB.Repository('fts_' + this.currentMode)
         .using('fts')
-        .where('', 'match', this.query)
+        .where('', 'match', qText)
       if (this.isPeriod && this.period) {
         repo = repo.where('ftsDate', '>=', this.period[0])
           .where('ftsDate', '<=', this.period[1])